Update frFR.lua
Not sure why the file here was empty? Work has been done before me, don't get me wrong. But it was a bit messy sometimes. So cleaned up, switched from alphabetical sorting to usUS file's structure for easier futur changes. Didn't do the changelog part tho as it is a lot of work and the few that actually read changelogs are usually able to understand it in english. :p